Change the way ratings' scores are calculated and stored
Previously, the numerical score was passed from the view, with most always = 3, and least always = 1. However, if the category is one in which a low score is better, it'd be easier to flip the numerical score, and 3 for least and 1 for most. That way, the overall score is easier to calculate: you can just add up the numerical scores stored, because the bad scores are already lower.